requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
pytest-asyncio
httpx
factory_boy
faker